painter/requirements.txt
pyautogui
pywin32
Pillow
PySide6
keyboard
numpy<2
pynput
